HOP-1904: move generic database connection to plugin
- remove greenplum repository
- move repositories to HTTPS
