+++ assemblies/plugins/engines/beam/pom.xml
@@ -46,7 +46,7 @@
     </repository>
     <repository>
       <id>confluent</id>
-      <url>http://packages.confluent.io/maven/</url>
+      <url>https://packages.confluent.io/maven/</url>
     </repository>
   </repositories>
