+++ core/src/main/java/org/apache/hop/core/database/DatabaseFactory.java
@@ -77,9 +77,9 @@ public class DatabaseFactory implements IDatabaseFactory {
   private StringBuilder appendConnectionInfo( IVariables variables, StringBuilder report, Database db, DatabaseMeta databaseMeta ) {
 
     // Check to see if the interface is of a type GenericDatabaseMeta, since it does not have hostname and port fields
-    if ( databaseMeta.getIDatabase() instanceof GenericDatabaseMeta ) {
+    if ( databaseMeta.getIDatabase().getPluginId() == "GENERIC" ) {
       
-    	String customDriverClass = databaseMeta.getAttributes().get( GenericDatabaseMeta.ATRRIBUTE_CUSTOM_DRIVER_CLASS );
+    	String customDriverClass = databaseMeta.getAttributes().get( NoneDatabaseMeta.ATRRIBUTE_CUSTOM_DRIVER_CLASS );
   		append( report, "GenericDatabaseMeta.report.customUrl",  db.resolve(  databaseMeta.getManualUrl() ) );
   		append( report, "GenericDatabaseMeta.report.customDriverClass",  db.resolve(  customDriverClass ) );
