+++ core/src/test/java/org/apache/hop/core/database/DatabaseTest.java
@@ -711,7 +711,7 @@ public class DatabaseTest {
     when( rsMetaData.getColumnType( 1 ) ).thenReturn( Types.DECIMAL );
 
     when( meta.stripCR( anyString() ) ).thenReturn( sql );
-    when( meta.getIDatabase() ).thenReturn( new GenericDatabaseMeta() );  // MySQL specific ?
+    when( meta.getIDatabase() ).thenReturn( new NoneDatabaseMeta() );  // MySQL specific ?
     when( conn.prepareStatement( sql, ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY ) ).thenReturn( ps );
     when( ps.getMetaData() ).thenReturn( rsMetaData );
 
@@ -733,7 +733,7 @@ public class DatabaseTest {
     when( ps.executeQuery() ).thenReturn( rs );
 
     when( meta.stripCR( anyString() ) ).thenReturn( sql );
-    when( meta.getIDatabase() ).thenReturn( new GenericDatabaseMeta() ); // MySQL specific ?
+    when( meta.getIDatabase() ).thenReturn( new NoneDatabaseMeta() ); // MySQL specific ?
     when( conn.prepareStatement( sql ) ).thenReturn( ps );
     when( rs.getMetaData() ).thenReturn( rsMetaData );
