+++ plugins/databases/generic/pom.xml
@@ -20,52 +20,35 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
     <parent>
+        <artifactId>hop-plugins-databases</artifactId>
         <groupId>org.apache.hop</groupId>
-        <artifactId>hop-assemblies-plugins-transforms</artifactId>
         <version>0.99-SNAPSHOT</version>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
 
+    <artifactId>hop-databases-generic</artifactId>
 
-    <artifactId>hop-assemblies-plugins-transforms-webservices</artifactId>
-    <version>0.99-SNAPSHOT</version>
-    <packaging>pom</packaging>
+    <packaging>jar</packaging>
 
-    <name>Hop Assemblies Plugins Transforms Web Services</name>
-    <description></description>
+    <name>Hop Plugins Databases Generic</name>
+
+    <properties>
+        <powermock.version>1.7.4</powermock.version>
+    </properties>
 
     <dependencies>
+        <!-- Test dependencies -->
         <dependency>
-            <groupId>org.apache.hop</groupId>
-            <artifactId>hop-transform-webservices</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.powermock</groupId>
+            <artifactId>powermock-module-junit4</artifactId>
+            <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>javax.wsdl</groupId>
-            <artifactId>com.springsource.javax.wsdl</artifactId>
-            <version>1.6.1</version>
+            <groupId>org.powermock</groupId>
+            <artifactId>powermock-api-mockito</artifactId>
+            <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>wsdl4j</groupId>
-            <artifactId>wsdl4j</artifactId>
-            <version>1.6.2</version>
-        </dependency>
+
     </dependencies>
-
-    <repositories>
-        <repository>
-            <id>EBIPublic</id>
-            <url>https://www.ebi.ac.uk/intact/maven/nexus/content/repositories/public/</url>
-            <name>EBIPublic Repository</name>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-            <releases>
-                <enabled>true</enabled>
-            </releases>
-        </repository>
-    </repositories>
-
 </project>
