+++ plugins/databases/greenplum/pom.xml
@@ -39,11 +39,6 @@
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>com.pivotal</groupId>
-      <artifactId>greenplum-jdbc</artifactId>
-      <version>${greenplum.version}</version>
-    </dependency>
   </dependencies>
 
   <packaging>jar</packaging>
@@ -99,11 +94,4 @@
     </plugins>
   </build>
 
-  <repositories>
-    <repository>
-      <id>pivotal</id>
-      <name>Pivotal Repository</name>
-      <url>http://nexus.saas.hand-china.com/content/repositories/rdc/</url>
-    </repository>
-  </repositories>
 </project>
