+++ plugins/databases/pom.xml
@@ -65,6 +65,7 @@
       </activation>
       <modules>
         <module>postgresql</module>
+        <module>generic</module>
         <module>greenplum</module>
         <module>redshift</module>
         <module>vertica</module>
