+++ plugins/transforms/execsqlrow/src/test/java/org/apache/hop/pipeline/transforms/execsqlrow/ExecSqlRowMetaInjectionTest.java
@@ -46,7 +46,7 @@ public class ExecSqlRowMetaInjectionTest extends BaseMetadataInjectionTest<ExecS
 
     skipPropertyTest( "CONNECTION_NAME" );
 
-    DatabaseMeta dbMeta = new DatabaseMeta( "testDBMeta", "Generic", "Native", "localhost", "test", "3306", "user", "password" );
+    DatabaseMeta dbMeta = new DatabaseMeta( "testDBMeta", "None", "Native", "localhost", "test", "3306", "user", "password" );
     metadataProvider.getSerializer( DatabaseMeta.class ).save(dbMeta);
     meta.setMetadataProvider( metadataProvider );
