+++ plugins/transforms/webservices/pom.xml
@@ -48,9 +48,9 @@
 
     <repositories>
         <repository>
-            <id>geomajas</id>
-            <url>http://maven.geomajas.org/</url>
-            <name>Geomajas</name>
+            <id>EBIPublic</id>
+            <url>https://www.ebi.ac.uk/intact/maven/nexus/content/repositories/public/</url>
+            <name>EBIPublic Repository</name>
             <snapshots>
                 <enabled>false</enabled>
             </snapshots>
